Move `/vault`
